+++ keywords/java/step-library-kw-selenium/src/main/java/ch/exense/step/examples/selenium/keyword/ChromeDriverKeyword.java
@@ -15,6 +15,7 @@
  ******************************************************************************/
 package ch.exense.step.examples.selenium.keyword;
 
+import ch.exense.step.examples.selenium.helper.AbstractSeleniumKeyword;
 import ch.exense.step.examples.selenium.helper.ProxyWrapper;
 import net.lightbody.bmp.BrowserMobProxy;
 import net.lightbody.bmp.BrowserMobProxyServer;
