+++ keywords/java/step-library-kw-selenium/src/main/java/ch/exense/step/examples/selenium/keyword/GenericSeleniumKeyword.java
@@ -16,6 +16,7 @@
 package ch.exense.step.examples.selenium.keyword;
 
 import ch.exense.step.examples.selenium.helper.AbstractPageObject;
+import ch.exense.step.examples.selenium.helper.AbstractSeleniumKeyword;
 import org.apache.commons.lang3.StringUtils;
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.ui.Select;
