+++ step-library-parent/pom.xml
@@ -82,6 +82,7 @@
 		<module>../keywords/java/step-library-kw-system</module>
 		<module>../keywords/java/step-library-kw-monitoring-system</module>
 		<module>../keywords/java/step-library-kw-step-client</module>
+		<module>../keywords/java/step-library-lib-selenium</module>
 		<module>../keywords/java/step-library-kw-selenium</module>
 		<module>../keywords/java/step-library-kw-excel</module>
 		<module>../keywords/java/step-library-kw-network</module>
